requirements.txt
altgraph==0.16.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
clcache==4.2.1.dev4+g0ab5d78
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
downtool==0.3.0
future==0.17.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
intel-openmp==2019.0
joblib==1.0.1
kiwisolver==1.2.0
macholib==1.11
matplotlib==3.3.2
mkl==2019.0
natsort==7.1.1
networkx==2.5
ninja==1.8.2.post2
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.5
pefile==2018.8.8
Pillow==8.0.0
protobuf==3.13.0
pycparser==2.20
PyInstaller==3.4
pymemcache==2.0.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyuv==1.4.0
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
PyYAML==3.13
requests==2.25.1
scikit-image==0.17.2
scipy==1.5.3
six==1.12.0
socketIO-client==0.7.2
tensorboardX==2.1
tifffile==2020.9.3
torch @ file:///E:/zjh/torch-1.6.0%2Bcu101-cp36-cp36m-win_amd64.whl
torchstat==0.0.7
torchsummary==1.5.1
torchvision @ file:///E:/zjh/torchvision-0.7.0%2Bcu101-cp36-cp36m-win_amd64.whl
tqdm==4.60.0
urllib3==1.26.4
warmup-scheduler==0.3
websocket-client==0.58.0
yacs==0.1.8
